Refactored the entire menu to be more modular and reusable
